🔨 [#053] - Split weekly-calendar and schedule-config-rows into focused components, fix e2e infrastructure 🧹
- 🗂️ Extracted WeekDayRow to week-day-row.tsx (from weekly-calendar.tsx)
- 🗂️ Extracted OverrideListDialog to override-list-dialog.tsx (from weekly-calendar.tsx)
- 🗂️ Extracted DayLabel to day-label.tsx (from schedule-config-rows.tsx)
- 🔧 Fixed cypress headless service: custom Dockerfile with Docker CLI + docker socket mount
- 🌐 Fixed e2e nginx: sushigo.e2e.local now proxies to Vite (port 5173) instead of API (port 80)
- 🌐 Added api.sushigo.e2e.local nginx block pointing to e2e Laravel API
- 🔧 Added VITE_API_URL=https://api.sushigo.e2e.local/api/v1 to test_e2e container
- ✅ Suppressed Vite HMR WebSocket uncaught exceptions in Cypress support file
- ✅ 23/23 e2e tests passing (employees.cy.ts + home.cy.ts + login.cy.ts)
