+++ docker/nginx/conf.d/sushigo.conf
@@ -328,10 +328,10 @@ server {
         return 204;
     }
 
-    # Try to proxy to e2e container, fallback to error page
+    # Try to proxy to e2e Vite dev server, fallback to error page
     location / {
         # Use variable for dynamic resolution - will fail gracefully if container doesn't exist
-        set $backend_e2e "test_e2e:80";
+        set $backend_e2e "test_e2e:5173";
         proxy_pass http://$backend_e2e;
         proxy_set_header Host $host;
         proxy_set_header X-Real-IP $remote_addr;
@@ -357,6 +357,77 @@ server {
 
     # Fallback page when e2e container is not available
     location @e2e_unavailable {
+        root /usr/share/nginx/html;
+        try_files /e2e-unavailable.html =503;
+    }
+}
+
+# api.sushigo.e2e.local - HTTP (redirect to HTTPS)
+server {
+    listen 80;
+    server_name api.sushigo.e2e.local;
+
+    location / {
+        return 301 https://$host$request_uri;
+    }
+}
+
+# api.sushigo.e2e.local - HTTPS (Laravel API on e2e container)
+server {
+    listen 443 ssl;
+    http2 on;
+    server_name api.sushigo.e2e.local;
+
+    ssl_certificate /etc/nginx/certs/sushigo.local.crt;
+    ssl_certificate_key /etc/nginx/certs/sushigo.local-key.pem;
+    ssl_protocols TLSv1.2 TLSv1.3;
+    ssl_prefer_server_ciphers on;
+    ssl_ciphers ECDHE-RSA-AES256-GCM-SHA512:DHE-RSA-AES256-GCM-SHA512:ECDHE-RSA-AES256-GCM-SHA384:DHE-RSA-AES256-GCM-SHA384;
+
+    # Security headers
+    add_header X-Frame-Options "SAMEORIGIN" always;
+    add_header X-Content-Type-Options "nosniff" always;
+    add_header X-XSS-Protection "1; mode=block" always;
+
+    # CORS headers - Allow e2e frontend with credentials
+    add_header Access-Control-Allow-Origin "https://sushigo.e2e.local" always;
+    add_header Access-Control-Allow-Methods "GET, POST, PUT, PATCH, DELETE, OPTIONS" always;
+    add_header Access-Control-Allow-Headers "Content-Type, Authorization, X-Requested-With, Accept, Origin" always;
+    add_header Access-Control-Allow-Credentials "true" always;
+    add_header Access-Control-Max-Age "3600" always;
+
+    # Handle preflight requests
+    if ($request_method = 'OPTIONS') {
+        return 204;
+    }
+
+    location / {
+        set $backend_e2e_api "test_e2e:80";
+        proxy_pass http://$backend_e2e_api;
+
+        # Hide duplicate CORS headers from backend
+        proxy_hide_header Access-Control-Allow-Origin;
+        proxy_hide_header Access-Control-Allow-Methods;
+        proxy_hide_header Access-Control-Allow-Headers;
+        proxy_hide_header Access-Control-Allow-Credentials;
+        proxy_hide_header Access-Control-Max-Age;
+
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
+        proxy_set_header X-Forwarded-Host $host;
+        proxy_set_header X-Forwarded-Port $server_port;
+
+        proxy_connect_timeout 5s;
+        proxy_send_timeout 60s;
+        proxy_read_timeout 60s;
+
+        proxy_intercept_errors on;
+        error_page 502 503 504 = @e2e_api_unavailable;
+    }
+
+    location @e2e_api_unavailable {
         root /usr/share/nginx/html;
         try_files /e2e-unavailable.html =503;
     }
